+++ src/app/features/home/activities/network-action-orders/network-action-orders.component.ts
@@ -1,11 +1,11 @@
 import { Component } from '@angular/core';
-import { UntilDestroy, untilDestroyed } from '@ngneat/until-destroy';
-import { BehaviorSubject } from 'rxjs';
-import { catchError, finalize } from 'rxjs/operators';
+import { UntilDestroy } from '@ngneat/until-destroy';
+import { BehaviorSubject, combineLatest } from 'rxjs';
+import { catchError, finalize, first, map, tap } from 'rxjs/operators';
 import {
-  BubbleOrderHistoryRecord,
-  OrderHistoryService,
-} from '../../../../shared/actions/service/order-history.service';
+  DiaBackendStoreService,
+  NetworkAppOrderWithThumbnail,
+} from '../../../../shared/dia-backend/store/dia-backend-store.service';
 import { ErrorService } from '../../../../shared/error/error.service';
 
 @UntilDestroy({ checkProperties: true })
@@ -15,30 +15,46 @@ import { ErrorService } from '../../../../shared/error/error.service';
   styleUrls: ['./network-action-orders.component.scss'],
 })
 export class NetworkActionOrdersComponent {
-  readonly networkActionOrders$ = this.orderHistoryService.networkActionOrders$;
-
+  readonly orders$ = new BehaviorSubject<NetworkAppOrderWithThumbnail[]>([]);
   readonly isFetching$ = new BehaviorSubject<boolean>(false);
 
+  readonly isOrdersEmpty$ = combineLatest([
+    this.isFetching$,
+    this.orders$,
+  ]).pipe(map(([isFetching, items]) => !isFetching && items.length === 0));
+
+  private readonly limit = 15;
+  private offset = 0;
+
   constructor(
-    private readonly orderHistoryService: OrderHistoryService,
+    private readonly storeService: DiaBackendStoreService,
     private readonly errorService: ErrorService
   ) {
+    this.loadData();
+  }
+
+  loadData(event?: any) {
     this.isFetching$.next(true);
-    this.orderHistoryService
-      .refresh$()
+    this.storeService
+      .listAllNetworkAppOrderWithThumbnail$({
+        offset: this.offset,
+        limit: this.limit,
+      })
       .pipe(
+        first(),
+        tap(({ results }) => {
+          this.orders$.next([...this.orders$.value, ...results]);
+          this.offset += this.limit;
+          event?.target?.complete();
+        }),
         catchError((err: unknown) => this.errorService.toastError$(err)),
-        finalize(() => this.isFetching$.next(false)),
-        untilDestroyed(this)
+        finalize(() => this.isFetching$.next(false))
       )
       .subscribe();
   }
 
   // eslint-disable-next-line class-methods-use-this
-  trackNetworkActionOrderHistoryRecords(
-    _: number,
-    item: BubbleOrderHistoryRecord
-  ) {
-    return item._id;
+  trackByNetworkActionOrder(_: number, item: NetworkAppOrderWithThumbnail) {
+    return item.id;
   }
 }