+++ src/app/features/home/capture-tab/capture-tab.component.ts
@@ -1,16 +1,25 @@
 import { formatDate, KeyValue } from '@angular/common';
 import { HttpErrorResponse } from '@angular/common/http';
-import { Component, OnInit } from '@angular/core';
-import { AlertController } from '@ionic/angular';
+import { ChangeDetectorRef, Component, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
+import { Router } from '@angular/router';
+import {
+  ActionSheetButton,
+  ActionSheetController,
+  AlertController,
+} from '@ionic/angular';
 import { TranslocoService } from '@ngneat/transloco';
 import { UntilDestroy, untilDestroyed } from '@ngneat/until-destroy';
 import { groupBy } from 'lodash-es';
-import { BehaviorSubject, combineLatest, iif } from 'rxjs';
+import { BehaviorSubject, combineLatest, defer, iif } from 'rxjs';
 import {
   catchError,
-  finalize,
+  concatMap,
+  concatMapTo,
   map,
   pluck,
+  shareReplay,
+  startWith,
   switchMap,
   tap,
 } from 'rxjs/operators';
@@ -19,17 +28,25 @@ import {
   CaptureTabSegments,
   CaptureTabService,
 } from '../../../shared/capture-tab/capture-tab.service';
+import { ConfirmAlert } from '../../../shared/confirm-alert/confirm-alert.service';
+import { Database } from '../../../shared/database/database.service';
 import {
   DiaBackendAsset,
   DiaBackendAssetRepository,
 } from '../../../shared/dia-backend/asset/dia-backend-asset-repository.service';
 import { DiaBackendAsseRefreshingService } from '../../../shared/dia-backend/asset/refreshing/dia-backend-asset-refreshing.service';
+import { DiaBackendAssetUploadingService } from '../../../shared/dia-backend/asset/uploading/dia-backend-asset-uploading.service';
 import { DiaBackendAuthService } from '../../../shared/dia-backend/auth/dia-backend-auth.service';
+import { DiaBackendTransactionRepository } from '../../../shared/dia-backend/transaction/dia-backend-transaction-repository.service';
 import { ErrorService } from '../../../shared/error/error.service';
+import { MediaStore } from '../../../shared/media/media-store/media-store.service';
 import { NetworkService } from '../../../shared/network/network.service';
+import { PreferenceManager } from '../../../shared/preference-manager/preference-manager.service';
 import { getOldProof } from '../../../shared/repositories/proof/old-proof-adapter';
 import { Proof } from '../../../shared/repositories/proof/proof';
 import { ProofRepository } from '../../../shared/repositories/proof/proof-repository.service';
+import { reloadApp } from '../../../utils/miscellaneous';
+import { PrefetchingDialogComponent } from '../onboarding/prefetching-dialog/prefetching-dialog.component';
 
 @UntilDestroy({ checkProperties: true })
 @Component({
@@ -45,9 +62,27 @@ export class CaptureTabComponent implements OnInit {
   readonly captureTabSegments = CaptureTabSegments;
   segment: CaptureTabSegments = CaptureTabSegments.COLLECTED;
 
+  readonly hasNewInbox$ = this.diaBackendTransactionRepository.inbox$.pipe(
+    catchError((err: unknown) => this.errorService.toastError$(err)),
+    map(transactions => transactions.count !== 0),
+    /**
+     * WORKARDOUND: force changeDetection to update badge when returning to App
+     * by clicking push notification
+     */
+    tap(() => this.changeDetectorRef.detectChanges()),
+    startWith(false)
+  );
+
   readonly username$ = this.diaBackendAuthService.username$;
 
   readonly email$ = this.diaBackendAuthService.email$;
+
+  readonly profileDescription$ = this.diaBackendAuthService.profileDescription$;
+
+  readonly profileBackground$ =
+    this.diaBackendAuthService.profileBackground$.pipe(
+      shareReplay({ bufferSize: 1, refCount: true })
+    );
 
   private readonly proofs$ = this.proofRepository.all$;
 
@@ -119,18 +154,34 @@ export class CaptureTabComponent implements OnInit {
     )
   );
 
+  private pendingUploadTasks = 0;
+
   constructor(
+    private readonly actionSheetController: ActionSheetController,
+    private readonly router: Router,
+    private readonly mediaStore: MediaStore,
+    private readonly database: Database,
+    private readonly confirmAlert: ConfirmAlert,
+    private readonly dialog: MatDialog,
+    private readonly preferenceManager: PreferenceManager,
+    private readonly changeDetectorRef: ChangeDetectorRef,
     private readonly proofRepository: ProofRepository,
     private readonly diaBackendAuthService: DiaBackendAuthService,
     private readonly diaBackendAssetRepository: DiaBackendAssetRepository,
+    private readonly diaBackendTransactionRepository: DiaBackendTransactionRepository,
     private readonly diaBackendAssetRefreshingService: DiaBackendAsseRefreshingService,
     private readonly alertController: AlertController,
     private readonly networkService: NetworkService,
     private readonly translocoService: TranslocoService,
     private readonly errorService: ErrorService,
     private readonly blockingActionService: BlockingActionService,
-    private readonly captureTabService: CaptureTabService
-  ) {}
+    private readonly captureTabService: CaptureTabService,
+    private readonly uploadService: DiaBackendAssetUploadingService
+  ) {
+    this.uploadService.pendingTasks$
+      .pipe(untilDestroyed(this))
+      .subscribe(value => (this.pendingUploadTasks = value));
+  }
 
   ngOnInit(): void {
     this.initSegmentListener();
@@ -159,6 +210,69 @@ export class CaptureTabComponent implements OnInit {
 
     const eventTarget = event.target as HTMLIonInfiniteScrollElement;
     eventTarget.complete();
+  }
+
+  openMenu() {
+    const buttons: ActionSheetButton[] = [
+      {
+        text: this.translocoService.translate('wallets.wallets'),
+        handler: () => this.router.navigate(['wallets']),
+      },
+      {
+        text: this.translocoService.translate('invitation.inviteFriends'),
+        handler: () => this.router.navigate(['invitation']),
+      },
+      {
+        text: this.translocoService.translate('settings.settings'),
+        handler: () => this.router.navigate(['settings']),
+      },
+      {
+        text: this.translocoService.translate('friends'),
+        handler: () => this.router.navigate(['contacts']),
+      },
+      {
+        text: this.translocoService.translate('dataPolicy'),
+        handler: () => this.router.navigate(['data-policy']),
+      },
+      {
+        text: this.translocoService.translate('termsOfUse'),
+        handler: () => this.router.navigate(['terms-of-use']),
+      },
+      {
+        text: this.translocoService.translate('faq'),
+        handler: () => this.router.navigate(['faq']),
+      },
+      {
+        text: this.translocoService.translate('logout'),
+        handler: () => this.logout(),
+        cssClass: 'capture-tab-menu-logout-button',
+      },
+    ];
+
+    this.actionSheetController
+      .create({ buttons })
+      .then(sheet => sheet.present());
+  }
+
+  logout() {
+    const action$ = defer(() => this.mediaStore.clear()).pipe(
+      concatMapTo(defer(() => this.database.clear())),
+      concatMapTo(defer(() => this.preferenceManager.clear())),
+      concatMapTo(defer(reloadApp)),
+      catchError((err: unknown) => this.errorService.toastError$(err))
+    );
+    defer(() =>
+      this.confirmAlert.present({
+        message: this.translocoService.translate('message.confirmLogout'),
+      })
+    )
+      .pipe(
+        concatMap(result =>
+          iif(() => result, this.blockingActionService.run$(action$))
+        ),
+        untilDestroyed(this)
+      )
+      .subscribe();
   }
 
   async editUsername() {
@@ -230,17 +344,32 @@ export class CaptureTabComponent implements OnInit {
     return item.id;
   }
 
-  refreshCaptures(event: Event) {
-    this.diaBackendAssetRefreshingService
-      .refresh()
-      .pipe(
-        finalize(() => {
-          this.capturedTabPageIndex$.next(0);
-          this.collectedTabPageIndex$.next(0);
-          this.draftTabPageIndex$.next(0);
-          return (<CustomEvent>event).detail.complete();
-        })
-      )
-      .subscribe();
+  async refreshCaptures(event: Event) {
+    (<CustomEvent>event).detail.complete();
+
+    // Don't refresh if there are still captures being uploaded.
+    if (this.pendingUploadTasks > 0) return;
+
+    const confirmRefresh = await this.showRefreshAlert();
+    if (confirmRefresh) {
+      this.capturedTabPageIndex$.next(0);
+      this.collectedTabPageIndex$.next(0);
+      this.draftTabPageIndex$.next(0);
+
+      return this.dialog.open(PrefetchingDialogComponent, {
+        disableClose: true,
+      });
+    }
+  }
+
+  private async showRefreshAlert() {
+    return this.confirmAlert.present({
+      header: this.translocoService.translate('syncAndRestore'),
+      message: this.translocoService.translate('message.confirmSyncAndRestore'),
+      confirmButtonText: this.translocoService.translate(
+        'confirmSyncAndRestore'
+      ),
+      cancelButtonText: this.translocoService.translate('cancelSyncAndRestore'),
+    });
   }
 }