+++ src/app/features/home/details/actions/action-details/action-details-routing.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+
+import { ActionDetailsPage } from './action-details.page';
+
+const routes: Routes = [
+  {
+    path: '',
+    component: ActionDetailsPage,
+  },
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule],
+})
+export class ActionDetailsPageRoutingModule {}